+++ src/models/json-api.model.ts
@@ -24,8 +24,7 @@ export class JsonApiModel {
 
   lastSyncModels: Array<any>;
 
-  // tslint:disable-next-line:variable-name
-  constructor(private _datastore: JsonApiDatastore, data?: any) {
+  constructor(private internalDatastore: JsonApiDatastore, data?: any) {
     if (data) {
       this.modelInitialization = true;
       this.id = data.id;
@@ -34,11 +33,11 @@ export class JsonApiModel {
     }
   }
 
-  isModelInitialization(): boolean {
+  public isModelInitialization(): boolean {
     return this.modelInitialization;
   }
 
-  syncRelationships(data: any, included: any, remainingModels?: Array<any>): void {
+  public syncRelationships(data: any, included: any, remainingModels?: Array<any>): void {
     if (this.lastSyncModels === included) {
       return;
     }
@@ -57,10 +56,10 @@ export class JsonApiModel {
     this.lastSyncModels = included;
   }
 
-  save(params?: any, headers?: HttpHeaders, customUrl?: string): Observable<this> {
+  public save(params?: any, headers?: HttpHeaders, customUrl?: string): Observable<this> {
     this.checkChanges();
     const attributesMetadata: any = this[AttributeMetadataIndex];
-    return this._datastore.saveRecord(attributesMetadata, this, params, headers, customUrl);
+    return this.internalDatastore.saveRecord(attributesMetadata, this, params, headers, customUrl);
   }
 
   get hasDirtyAttributes() {
@@ -99,7 +98,7 @@ export class JsonApiModel {
     }
   }
 
-  rollbackAttributes(): void {
+  public rollbackAttributes(): void {
     const attributesMetadata: any = this[AttributeMetadataIndex];
     for (const propertyName in attributesMetadata) {
       if (attributesMetadata.hasOwnProperty(propertyName)) {
@@ -113,7 +112,6 @@ export class JsonApiModel {
   get modelConfig(): ModelConfig {
     return Reflect.getMetadata('JsonApiModelConfig', this.constructor);
   }
-
 
   private parseHasMany(data: any, included: any, remainingModels: Array<any>): void {
     const hasMany: any = Reflect.getMetadata('HasMany', this);
@@ -132,7 +130,7 @@ export class JsonApiModel {
             if (!includes(modelTypesFetched, typeName)) {
               modelTypesFetched.push(typeName);
               // tslint:disable-next-line:max-line-length
-              const modelType: ModelType<this> = Reflect.getMetadata('JsonApiDatastoreConfig', this._datastore.constructor).models[typeName];
+              const modelType: ModelType<this> = Reflect.getMetadata('JsonApiDatastoreConfig', this.internalDatastore.constructor).models[typeName];
 
               if (modelType) {
                 const relationshipModels: JsonApiModel[] = this.getHasManyRelationship(
@@ -147,7 +145,7 @@ export class JsonApiModel {
                   allModels = allModels.concat(relationshipModels);
                 }
               } else {
-                throw { message: 'parseHasMany - Model type for relationship ' + typeName + ' not found.' };
+                throw { message: `parseHasMany - Model type for relationship ${typeName} not found.` };
               }
             }
 
@@ -171,7 +169,7 @@ export class JsonApiModel {
           if (dataRelationship) {
             const typeName: string = dataRelationship.type;
             // tslint:disable-next-line:max-line-length
-            const modelType: ModelType<this> = Reflect.getMetadata('JsonApiDatastoreConfig', this._datastore.constructor).models[typeName];
+            const modelType: ModelType<this> = Reflect.getMetadata('JsonApiDatastoreConfig', this.internalDatastore.constructor).models[typeName];
 
             if (modelType) {
               const relationshipModel = this.getBelongsToRelationship(
@@ -186,7 +184,7 @@ export class JsonApiModel {
                 this[metadata.propertyName] = relationshipModel;
               }
             } else {
-              throw { message: 'parseBelongsTo - Model type for relationship ' + typeName + ' not found.' };
+              throw { message: `parseBelongsTo - Model type for relationship ${typeName} not found.` };
             }
           }
         }
@@ -204,7 +202,7 @@ export class JsonApiModel {
     const relationshipList: Array<T> = [];
 
     data.forEach((item: any) => {
-      const relationshipData: any = find(remainingModels, { id: item.id, type: typeName });
+      const relationshipData: any = find(remainingModels, { id: item.id, type: typeName } as any);
 
       if (relationshipData) {
         const newObject: T = this.createOrPeek(modelType, relationshipData);
@@ -222,7 +220,6 @@ export class JsonApiModel {
     return relationshipList;
   }
 
-
   private getBelongsToRelationship<T extends JsonApiModel>(
     modelType: ModelType<T>,
     data: any,
@@ -232,7 +229,7 @@ export class JsonApiModel {
   ): T | null {
     const id: string = data.id;
 
-    const relationshipData: any = find(remainingModels, { id, type: typeName });
+    const relationshipData: any = find(remainingModels, { id, type: typeName } as any);
 
     if (relationshipData) {
       const newObject: T = this.createOrPeek(modelType, relationshipData);
@@ -246,19 +243,19 @@ export class JsonApiModel {
       return newObject;
     }
 
-    return this._datastore.peekRecord(modelType, id);
+    return this.internalDatastore.peekRecord(modelType, id);
   }
 
   private createOrPeek<T extends JsonApiModel>(modelType: ModelType<T>, data: any): T {
-    const peek = this._datastore.peekRecord(modelType, data.id);
+    const peek = this.internalDatastore.peekRecord(modelType, data.id);
 
     if (peek) {
-      _.extend(peek, this._datastore.transformSerializedNamesToPropertyNames(modelType, data.attributes));
+      _.extend(peek, this.internalDatastore.transformSerializedNamesToPropertyNames(modelType, data.attributes));
       return peek;
     }
 
-    const newObject: T = this._datastore.deserializeModel(modelType, data);
-    this._datastore.addToStore(newObject);
+    const newObject: T = this.internalDatastore.deserializeModel(modelType, data);
+    this.internalDatastore.addToStore(newObject);
 
     return newObject;
   }