+++ src/services/json-api-datastore.service.ts
@@ -24,14 +24,12 @@ const AttributeMetadataIndex: string = AttributeMetadata as any;
 
 @Injectable()
 export class JsonApiDatastore {
-  // tslint:disable-next-line:variable-name
-  private _headers: HttpHeaders;
-  // tslint:disable-next-line:variable-name
-  private _store: {[type: string]: {[id: string]: JsonApiModel}} = {};
+  private globalHeaders: HttpHeaders;
+  private globalRequestOptions: object = {};
+  private internalStore: {[type: string]: {[id: string]: JsonApiModel}} = {};
   private toQueryString: Function = this.datastoreConfig.overrides
     && this.datastoreConfig.overrides.toQueryString ?
       this.datastoreConfig.overrides.toQueryString : this._toQueryString;
-  // tslint:enable:max-line-length
 
   private get getDirtyAttributes() {
     if (this.datastoreConfig.overrides
@@ -52,7 +50,7 @@ export class JsonApiDatastore {
     headers?: HttpHeaders,
     customUrl?: string
   ): Observable<T[]> {
-    const requestHeaders: HttpHeaders = this.buildHeaders(headers);
+    const requestHeaders: HttpHeaders = this.buildHttpHeaders(headers);
     const url: string = this.buildUrl(modelType, params, undefined, customUrl);
     return this.http.get(url, { headers: requestHeaders })
       .pipe(
@@ -61,40 +59,40 @@ export class JsonApiDatastore {
       );
   }
 
-  findAll<T extends JsonApiModel>(
+  public findAll<T extends JsonApiModel>(
     modelType: ModelType<T>,
     params?: any,
     headers?: HttpHeaders,
     customUrl?: string
   ): Observable<JsonApiQueryData<T>> {
-    const requestHeaders: HttpHeaders = this.buildHeaders(headers);
     const url: string = this.buildUrl(modelType, params, undefined, customUrl);
+    const requestOptions: object = this.buildRequestOptions({ headers, observe: 'response' });
 
-    return this.http.get(url, { headers: requestHeaders })
+    return this.http.get(url, requestOptions)
       .pipe(
-        map((res: any) => this.extractQueryData(res, modelType, true)),
+        map((res: HttpResponse<object>) => this.extractQueryData(res, modelType, true)),
         catchError((res: any) => this.handleError(res))
       );
   }
 
-  findRecord<T extends JsonApiModel>(
+  public findRecord<T extends JsonApiModel>(
     modelType: ModelType<T>,
     id: string,
     params?: any,
     headers?: HttpHeaders,
     customUrl?: string
   ): Observable<T> {
-    const requestHeaders: HttpHeaders = this.buildHeaders(headers);
+    const requestOptions: object = this.buildRequestOptions({ headers, observe: 'response' });
     const url: string = this.buildUrl(modelType, params, id, customUrl);
 
-    return this.http.get(url, { headers: requestHeaders, observe: 'response' })
+    return this.http.get(url, requestOptions)
       .pipe(
-        map((res) => this.extractRecordData(res, modelType)),
+        map((res: HttpResponse<object>) => this.extractRecordData(res, modelType)),
         catchError((res: any) => this.handleError(res))
       );
   }
 
-  createRecord<T extends JsonApiModel>(modelType: ModelType<T>, data?: any): T {
+  public createRecord<T extends JsonApiModel>(modelType: ModelType<T>, data?: any): T {
     return new modelType(this, { attributes: data });
   }
 
@@ -114,7 +112,7 @@ export class JsonApiDatastore {
     return dirtyData;
   }
 
-  saveRecord<T extends JsonApiModel>(
+  public saveRecord<T extends JsonApiModel>(
     attributesMetadata: any,
     model: T,
     params?: any,
@@ -124,11 +122,10 @@ export class JsonApiDatastore {
     const modelType = <ModelType<T>>model.constructor;
     const modelConfig: ModelConfig = model.modelConfig;
     const typeName: string = modelConfig.type;
-    const requestHeaders: HttpHeaders = this.buildHeaders(headers);
     const relationships: any = this.getRelationships(model);
     const url: string = this.buildUrl(modelType, params, model.id, customUrl);
 
-    let httpCall: Observable<HttpResponse<Object>>;
+    let httpCall: Observable<HttpResponse<object>>;
     const body: any = {
       data: {
         relationships,
@@ -138,10 +135,12 @@ export class JsonApiDatastore {
       }
     };
 
+    const requestOptions: object = this.buildRequestOptions({ headers, observe: 'response' });
+
     if (model.id) {
-      httpCall = this.http.patch(url, body, { headers: requestHeaders, observe: 'response' });
+      httpCall = this.http.patch<object>(url, body, requestOptions) as Observable<HttpResponse<object>>;
     } else {
-      httpCall = this.http.post(url, body, { headers: requestHeaders, observe: 'response' });
+      httpCall = this.http.post<object>(url, body, requestOptions) as Observable<HttpResponse<object>>;
     }
 
     return httpCall
@@ -153,39 +152,42 @@ export class JsonApiDatastore {
           }
           return this.handleError(res);
         }),
-        // map((res) => this.resetMetadataAttributes(res, attributesMetadata, modelType)),
         map((res) => this.updateRelationships(res, relationships))
       );
   }
 
-  deleteRecord<T extends JsonApiModel>(
+  public deleteRecord<T extends JsonApiModel>(
     modelType: ModelType<T>,
     id: string,
     headers?: HttpHeaders,
     customUrl?: string
   ): Observable<Response> {
-    const requestHeaders: HttpHeaders = this.buildHeaders(headers);
+    const requestOptions: object = this.buildRequestOptions({ headers });
     const url: string = this.buildUrl(modelType, null, id, customUrl);
 
-    return this.http.delete(url, { headers: requestHeaders })
+    return this.http.delete(url, requestOptions)
       .pipe(
         catchError((res: HttpErrorResponse) => this.handleError(res))
       );
   }
 
-  peekRecord<T extends JsonApiModel>(modelType: ModelType<T>, id: string): T | null {
+  public peekRecord<T extends JsonApiModel>(modelType: ModelType<T>, id: string): T | null {
     const type: string = Reflect.getMetadata('JsonApiModelConfig', modelType).type;
-    return this._store[type] ? <T>this._store[type][id] : null;
+    return this.internalStore[type] ? <T>this.internalStore[type][id] : null;
   }
 
-  peekAll<T extends JsonApiModel>(modelType: ModelType<T>): T[] {
+  public peekAll<T extends JsonApiModel>(modelType: ModelType<T>): Array<T> {
     const type = Reflect.getMetadata('JsonApiModelConfig', modelType).type;
-    const typeStore = this._store[type];
+    const typeStore = this.internalStore[type];
     return typeStore ? Object.keys(typeStore).map((key) => <T>typeStore[key]) : [];
   }
 
   set headers(headers: HttpHeaders) {
-    this._headers = headers;
+    this.globalHeaders = headers;
+  }
+
+  set requestOptions(requestOptions: object) {
+    this.globalRequestOptions = requestOptions;
   }
 
   protected buildUrl<T extends JsonApiModel>(
@@ -244,9 +246,13 @@ export class JsonApiDatastore {
 
   protected isValidToManyRelation(objects: Array<any>): boolean {
     const isJsonApiModel = objects.every((item) => item instanceof JsonApiModel);
-    const relationshipType: string = isJsonApiModel ? objects[0].modelConfig.type : '';
-
-    return isJsonApiModel ? objects.every((item: JsonApiModel) => item.modelConfig.type === relationshipType) : false;
+    if (!isJsonApiModel) {
+      return false;
+    }
+    const types = objects.map((item: JsonApiModel) => item.modelConfig.modelEndpointUrl || item.modelConfig.type);
+    return types
+      .filter((type: string, index: number, self: string[]) => self.indexOf(type) === index)
+      .length === 1;
   }
 
   protected buildSingleRelationshipData(model: JsonApiModel): any {
@@ -264,10 +270,11 @@ export class JsonApiDatastore {
   }
 
   protected extractQueryData<T extends JsonApiModel>(
-    body: any,
+    response: HttpResponse<object>,
     modelType: ModelType<T>,
     withMeta = false
-  ): T[] | JsonApiQueryData<T> {
+  ): Array<T> | JsonApiQueryData<T> {
+    const body: any = response.body;
     const models: T[] = [];
 
     body.data.forEach((data: any) => {
@@ -285,6 +292,7 @@ export class JsonApiDatastore {
     if (withMeta && withMeta === true) {
       return new JsonApiQueryData(models, this.parseMeta(body, modelType));
     }
+
     return models;
   }
 
@@ -331,7 +339,6 @@ export class JsonApiDatastore {
   }
 
   protected handleError(error: any): Observable<any> {
-
     if (
       error instanceof HttpErrorResponse &&
       error.error instanceof Object &&
@@ -339,11 +346,9 @@ export class JsonApiDatastore {
       error.error.errors instanceof Array
     ) {
       const errors: ErrorResponse = new ErrorResponse(error.error.errors);
-      console.error(error, errors);
       return throwError(errors);
     }
 
-    console.error(error);
     return throwError(error);
   }
 
@@ -352,28 +357,26 @@ export class JsonApiDatastore {
     return new metaModel(body);
   }
 
-  /** @deprecated - use buildHeaders method to build request headers **/
+  /** @deprecated - use buildHttpHeaders method to build request headers **/
   protected getOptions(customHeaders?: HttpHeaders): any {
     return {
-      headers: this.buildHeaders(customHeaders),
+      headers: this.buildHttpHeaders(customHeaders),
     };
   }
 
-  protected buildHeaders(customHeaders?: HttpHeaders): HttpHeaders {
-
+  protected buildHttpHeaders(customHeaders?: HttpHeaders): HttpHeaders {
     let requestHeaders: HttpHeaders = new HttpHeaders({
       Accept: 'application/vnd.api+json',
       'Content-Type': 'application/vnd.api+json'
     });
 
-    if (this._headers) {
-      this._headers.keys().forEach((key) => {
-        if (this._headers.has(key)) {
-          requestHeaders = requestHeaders.set(key, this._headers.get(key)!);
+    if (this.globalHeaders) {
+      this.globalHeaders.keys().forEach((key) => {
+        if (this.globalHeaders.has(key)) {
+          requestHeaders = requestHeaders.set(key, this.globalHeaders.get(key)!);
         }
       });
     }
-
 
     if (customHeaders) {
       customHeaders.keys().forEach((key) => {
@@ -386,6 +389,16 @@ export class JsonApiDatastore {
     return requestHeaders;
   }
 
+  private buildRequestOptions(customOptions: any = {}): object {
+    const httpHeaders: HttpHeaders = this.buildHttpHeaders(customOptions.headers);
+
+    const requestOptions: object = Object.assign(customOptions, {
+      headers: httpHeaders
+    });
+
+    return Object.assign(this.globalRequestOptions, requestOptions);
+  }
+
   private _toQueryString(params: any): string {
     return qs.stringify(params, { arrayFormat: 'brackets' });
   }
@@ -393,10 +406,10 @@ export class JsonApiDatastore {
   public addToStore(modelOrModels: JsonApiModel | JsonApiModel[]): void {
     const models = Array.isArray(modelOrModels) ? modelOrModels : [modelOrModels];
     const type: string = models[0].modelConfig.type;
-    let typeStore = this._store[type];
+    let typeStore = this.internalStore[type];
 
     if (!typeStore) {
-      typeStore = this._store[type] = {};
+      typeStore = this.internalStore[type] = {};
     }
 
     for (const model of models) {
@@ -405,7 +418,6 @@ export class JsonApiDatastore {
   }
 
   protected resetMetadataAttributes<T extends JsonApiModel>(res: T, attributesMetadata: any, modelType: ModelType<T>) {
-    // TODO check why is attributesMetadata from the arguments never used
     for (const propertyName in attributesMetadata) {
       if (attributesMetadata.hasOwnProperty(propertyName)) {
         const metadata: any = attributesMetadata[propertyName];
@@ -448,7 +460,7 @@ export class JsonApiDatastore {
     return model;
   }
 
-  protected get datastoreConfig(): DatastoreConfig {
+  public get datastoreConfig(): DatastoreConfig {
     const configFromDecorator: DatastoreConfig = Reflect.getMetadata('JsonApiDatastoreConfig', this.constructor);
     return Object.assign(configFromDecorator, this.config);
   }
@@ -456,6 +468,7 @@ export class JsonApiDatastore {
   public transformSerializedNamesToPropertyNames<T extends JsonApiModel>(modelType: ModelType<T>, attributes: any) {
     const serializedNameToPropertyName = this.getModelPropertyNames(modelType.prototype);
     const properties: any = {};
+
     Object.keys(serializedNameToPropertyName).forEach((serializedName) => {
       if (attributes && attributes[serializedName] !== null && attributes[serializedName] !== undefined) {
         properties[serializedNameToPropertyName[serializedName]] = attributes[serializedName];